feat!: Following refactor!: Refactored non-folder designated icons to designated folder
