+++ components/icons/GeneralIcons/SettingsIcon.tsx
