+++ components/icons/GeneralIcons/SignOutIcon.tsx
